pyproject.toml
[project]
name = "lumberjack-client"
description = "A simple client for a simple log management system."
readme = "README.md"
license = { text = "MIT" }
authors = [{name = "Michael de Villiers", email = "michael@devilears.co.za"},]
maintainers = [{name = "Michael de Villiers", email = "michael@devilears.co.za"},]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
]
dynamic = ["version"]
requires-python = ">= 3.7"
dependencies = [
    "click ~= 8.1.3",
    "python-dateutil ~= 2.8.2",
    "requests ~= 2.28.2",
    "requests_toolbelt ~= 0.10.1",
    "watchdog ~= 2.2.1",
]

[project.scripts]
jack = "lumberjack_client:main"

[project.urls]
Documentation = "https://lumberjack.readthedocs.io/client/"
Changes = "https://lumberjack.readthedocs.io/changes/"
"Source Code" = "https://github.com/COUR4G3/lumberjack-client/"
"Issue Tracker" = "https://github.com/COUR4G3/lumberjack-client/issues/"

[project.optional-dependencies]
dev = [
    "bandit",
    "black",
    "flake8",
    "pre_commit ~= 2.21.0",
    "pydocstyle",
    "reorder-python-imports",
    "lumberjack[test]",
]
test = [
    "pytest ~= 7.2.0",
    "pytest-cov ~= 4.0.0",
    "pytest-mock ~= 3.10.0",
    "pytest-skip-slow == 0.0.3"
]

[build-system]
requires = ["setuptools >= 64.0", "setuptools_scm[toml] >= 6.2", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
force-exclude = '''
(
  /venv/
)
'''
line-length = 79
target-version = ['py37']

[tool.coverage.run]
source = ["lumberjack_client"]

[tool.pytest.ini_options]
addopts = "--cov=lumberjack-client.py --cov-report html --cov-report term-missing"
markers = [
    "slow: tests that are slow",
    "integration: integration tests that require external tools",
]

[tool.setuptools_scm]
write_to = "_version.py"

